+++ test/endpoint.js
@@ -47,6 +47,7 @@ describe('phonetics', function() {
 describe('stats', function() {
   require('./stats/descriptive');
   require('./stats/frequencies.js');
+  require('./stats/inferential.js');
   require('./stats/ngrams.js');
   require('./stats/tfidf.js');
 });